+++ instrumentation/opentelemetry-instrumentation-sqlalchemy/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Update sqlalchemy instrumentation to follow semantic conventions
+([#202](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/202))
+
 ## Version 0.13b0
 
 Released 2020-09-17